Trotterization time evolution simulation and correlation function measurement
The following code can be used to measure two-point spin-spin dynamical correlation functions with the use of Trotterization as the time evolution simulation method on IBM quantum hardware. A standard Hadamard test (i.e., indirect measurements) is used to implement the measurements of the circuits. Plots of the real/imaginary parts of the correlation function vs. time as well as the Fourier transform vs. frequency are produced with the code.